Forcing explicit conversion to pass Clang

+++ beluga/include/beluga/sensor/likelihood_field_model_base.hpp
@@ -157,7 +157,7 @@ class LikelihoodFieldModelBase {
       std::vector<bool> effective_unknown_mask{std::begin(grid.unknown_mask()), std::end(grid.unknown_mask())};
 
       for (std::size_t idx = 0; idx < grid.size(); ++idx) {
-        if (!grid.obstacle_mask()[idx])
+        if (!grid.obstacle_mask()[static_cast<long>(idx)])
           continue;  // skip free cells
 
         // Check if any 4-neighbor is NOT an obstacle (free or unknown)
